# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayMyDine Restaurant API
  description: |
    Multi-tenant restaurant ordering and management system API.
    
    **Security Warning**: This API currently has no authentication implemented.
    All endpoints are publicly accessible.
    
    **Multi-tenant**: Each restaurant has isolated data through subdomain-based routing.
  version: 1.0.0
  contact:
    name: PayMyDine Support
    email: support@paymydine.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://{tenant}.paymydine.com/api/v1
    description: Production server (tenant-specific)
security:
  - {} # No authentication currently implemented
paths:
...
  /menu:
    get:
      summary: Get menu items
      description: Retrieve all available menu items with categories
      operationId: getMenu
      tags:
        - Menu
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /menu/categories/{categoryId}/items:
    get:
      summary: Get menu items by category
      description: Retrieve menu items for a specific category
      operationId: getMenuItemsByCategory
      tags:
        - Menu
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Category ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Menu items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
...
  /categories/{categoryId}:
    get:
      summary: Get category details
      description: Retrieve details for a specific category
      operationId: getCategory
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Category ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Category'
  /orders:
    post:
      summary: Create order
      description: Create a new order
      operationId: createOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  order_id:
                    type: integer
                    example: 123
                  message:
                    type: string
                    example: Order placed successfully
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List orders
      description: Retrieve list of orders
      operationId: listOrders
      tags:
        - Orders
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      summary: Get order details
      description: Retrieve details for a specific order
      operationId: getOrder
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update order
      description: Update an existing order
      operationId: updateOrder
      tags:
        - Orders
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order updated successfully
  /order-status:
    get:
      summary: Get order status
      description: Retrieve status for a specific order
      operationId: getOrderStatus
      tags:
        - Orders
      parameters:
        - name: order_id
          in: query
          required: true
          description: Order ID
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Order status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OrderStatus'

    post:
      summary: Update order status
      description: Update status for a specific order
      operationId: updateOrderStatus
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                  example: 123
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, delivered, cancelled]
                  example: preparing
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Order status updated successfully
...
  /tables/{qrCode}:
    get:
      summary: Get table by QR code
      description: Retrieve table information using QR code
      operationId: getTableByQrCode
      tags:
        - Tables
      parameters:
        - name: qrCode
          in: path
          required: true
          description: QR code
          schema:
            type: string
            example: ms10zQpxP
      responses:
        '200':
          description: Table retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Table'
        '404':
          description: Table not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /table-info:
    get:
      summary: Get table information
      description: Retrieve table information by table ID
      operationId: getTableInfo
      tags:
        - Tables
      parameters:
        - name: table_id
          in: query
          required: true
          description: Table ID
          schema:
            type: string
            example: "7"
      responses:
        '200':
          description: Table information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Table'
  /table-menu:
    get:
      summary: Get table menu
      description: Retrieve menu for a specific table
      operationId: getTableMenu
      tags:
        - Tables
      parameters:
        - name: table_id
          in: query
          required: true
          description: Table ID
          schema:
            type: string
            example: "7"
        - name: location_id
          in: query
          required: false
          description: Location ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Table menu retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/MenuItem'
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
...
  /waiter-call:
    post:
      summary: Call waiter
      description: Send a waiter call request
      operationId: callWaiter
      tags:
        - Restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table_id:
                  type: string
                  example: "7"
                message:
                  type: string
                  maxLength: 500
                  example: "Need assistance with the menu"
      responses:
        '200':
          description: Waiter call sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Waiter called successfully
                  call_id:
                    type: string
                    example: "64f8a1b2c3d4e5f6"
  /api/v1/valet-request:
    post:
      summary: Submit valet request
      description: Submit a valet parking request for a table
      operationId: submitValetRequest
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - table_id
                - customer_name
                - car_make
                - license_plate
              properties:
                table_id:
                  type: string
                  example: "7"
                customer_name:
                  type: string
                  maxLength: 255
                  example: "John Smith"
                car_make:
                  type: string
                  maxLength: 255
                  example: "Toyota Camry"
                license_plate:
                  type: string
                  maxLength: 20
                  example: "ABC123"
      responses:
        '200':
          description: Valet request submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Valet request submitted successfully"
                  request_id:
                    type: integer
                    example: 123
                  notification_id:
                    type: integer
                    example: 456
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Table not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Margherita Pizza"
        description:
          type: string
          example: "Classic pizza with tomato and mozzarella"
        price:
          type: number
          format: float
          example: 15.99
        image:
          type: string
          example: "/images/pizza.jpg"
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: "Pizza"
        calories:
          type: integer
          example: 300
        allergens:
          type: array
          items:
            type: string
          example: ["gluten", "dairy"]
        stock_qty:
          type: integer
          example: 10
        minimum_qty:
          type: integer
          example: 1
        available:
          type: boolean
          example: true
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Pizza"
        description:
          type: string
          example: "Italian pizzas"
        image:
          type: string
          example: "/images/pizza-category.jpg"
...
    Order:
      type: object
      properties:
        order_id:
          type: integer
          example: 123
        customer_id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        telephone:
          type: string
          example: "+1234567890"
        location_id:
          type: integer
          example: 1
        order_type:
          type: string
          example: "Table 7"
        order_total:
          type: number
          format: float
          example: 31.98
        status_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Order not found"
        message:
          type: string
          example: "The requested order does not exist"
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "The customer name field is required"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            customer_name: ["The customer name field is required"]
            items: ["The items field is required"]
...